requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.94.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
packaging==23.1
plotly==5.14.1
protobuf==3.20.3
pydantic==1.10.6
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
tenacity==8.2.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
